Separate implementation implementation :) into its own file.
